Rate and payment

+++ WAPIProject/DTO/ComputerDTO.cs
@@ -9,8 +9,8 @@ namespace WAPIProject.DTO
         public string Name { get; set; }
         public string BrandName { get; set; }
         public string Description { get; set; }
-        public decimal Price { get; set; }
-        public decimal? PriceAfterDiscount { get; set; }
+        public double Price { get; set; }
+        public double? PriceAfterDiscount { get; set; }
         public int Quantity { get; set; }
         public Stars RateValue { get; set; }
         public double Weight { get; set; }
